remove develpatch version number in generic manner
I was getting an error because we were trying to fetch an RStudio
verison that was incorrect because RStudio likes to publish version
numbers such as

1.2.5001-1

but the installers to be downloaded would not have the -1 suffix.,

In my madness last time I fixed this, I had assumed that the final
number would always be "1", but instead found out that it can be
anything, really. This will fix that issue.