fix length > 1 error for coercion to logical
If I had bothered to look at the failure logs on travis, I would have
found it: https://travis-ci.org/reconhub/nomad/builds/602254632#L1795

It turns out that the `target` parameter could have multiple targets, so
this would return a vector of length > 1.

My fix was to wrap it in `any()` since we are explicitly testing if any
of the targets include windows.